Add unit test for glsl-construct-vec-mat-index's expectations.
Add the ability to override the browser_type of the fake browser in
telemetry.testing.fakes.

The unit test is working as expected. Need to run this test repeatedly
on Windows to understand whether the test expectation isn't being
picked up, or whether the flaky test handling is unable to recover
from the kind of failure that's happening.

BUG=534697

Review URL: https://codereview.chromium.org/1358943002

Cr-Commit-Position: refs/heads/master@{#350254}