Fix wrapping behavior of rail_score_span
The railscore_span was set to display with inline, which isn't a display mode.
This caused it to default to flex.

The intent of this span is that it never become multiline. But, when the width
was too narrow, the spans it contained would shrink and word-wrapping would
ocurr. This caused the span to become multiline, which would then cause the UI
of the embedder to get confused.

Finally, the railscore vertical alignment was off. This made it look awkwardly
positioned when inside an otherwise-centered row of text.

This fixes these issues in straightforward ways (inasmuch as flex and css
fixes are ever straightforward).

TBR=benjhayden@chromium.org

Review URL: https://codereview.chromium.org/1632483002