Fix tracing's interaction with glMatix when used in node
Tracing's codebase currently assumes that all loading happens module-less: that is, that all imports go into the global scope. Whereas, glMatrix is smart enough to see that it is being brought into a Node module, so instead
of writing its exports into the global scope, it writes into its exports.

This then breaks tracing because tracing is still looking for glMatrix
in globals.

The fix for this is to, in node only, find the glMatix module, and copy
its exports onto the global object. Its ugly, but it'll work until we can
move tracing/ to be a proper module from the beginning.

TBR=nednguyen
CC=aerotwist,paulirish

Review URL: https://codereview.chromium.org/1668793002