Overhaul attribute info handling in Memory-Infra
This patch modifies Trace Viewer to generate attribute infos in the UI
instead of the model. The rationale for this is threefold:

  1. Performance. The new approach avoids high memory usage caused by
     lots of long strings stored in the model. Consequently, the V8
     heap sharnk by 27.5% (399 -> 289 MiB) after importing an 8-second
     trace from YouTube TV (see
     https://github.com/catapult-project/catapult/issues/1639).

  2. Software design. Since the strings are meant to be seen by the
     user, they should arguably be generated in the UI.

  3. Aggregation. The new framework makes it considerably easier to
     aggregate infos of multiple dumps in multi-selection mode.
     Consequently, all infos are now aggregated regardless of the
     number of selected dumps.

The patch is a huge refactor of existing code and will therefore remain
invisible to the end user (except for reduced memory usage and newly
added info aggregation in multi-selection mode).

BUG=catapult:#1883,catapult:#1639

Review URL: https://codereview.chromium.org/1629423004