Fix flaky test in get_logs.
This is a case when test runs too fast and new log Records get added with the same timestamp.  Same goes for live environment, but minor edge case.  Another solution is to get Records with equal or greater timestamp, and front-end remove the duplicates, but it's not a big deal.

Review URL: https://codereview.chromium.org/1721433002