[memory-infra] Move 'infos' field from Attribute to MemoryAllocatorDump
We want to switch from the memory-infra-specific Attribute class to the
Catapult-wide tr.v.ScalarNumeric class in the near future. The latter
does not support infos and the feature is currently only used for
warnings about the 'size' attribute anyway, so we're moving infos to the
associated MemoryAllocatorDump:

  dump.attrs['size'].infos ----> dump.infos

This also allows us to remove the unnecessary attribute cloning methods
(the list of infos used to be modified in the UI) since attributes are
now treated as immutable (just like numerics).

BUG=catapult:#1928

Review URL: https://codereview.chromium.org/1744563002