Un-register SystemHealthMetrics until 2098
Currently, importing a metric is conflated with specifying that it should be
run. And TBMv2 allows benchmarks to specify exactly one metric.

Eventually, we want a SystemHealthBenchmark that specifies that the metric to
run is "SystemHealthMetrics", and the SystemHealthMetrics metric should run
ResponsivenessMetric, EfficiencyMetric, etc.

While importing a metric is conflated with running it, this means that
SystemHealthMetrics cannot be a registered metric, but it must be a file that
imports the other metrics, so that they can register themselves.

BUG=catapult:#2098
R=eakuefner@chromium.org

Review URL: https://codereview.chromium.org/1803853002 .