Display per-UE metric values in single-user-expectation-sub-view
The UE sub-view used to hard-code calls to the system health metrics.
That was nice for debugging, but too special-case-y.

Currently, metrics are not displayed in trace viewer anywhere.

This CL computes all of the metrics whenever a UE is selected,
and displays only the values that pertain to the selected UE.

A full metric debugger grouping table is left for another day.

BUG=catapult:#1855

Review URL: https://codereview.chromium.org/1816353002