[clock sync] Make ftrace importer use a function to shift timestamps
In doing this, I also beefed up the documentation in
FTraceImporter.computeTimeTransform_().

The reason for using a function to shift timestamps instead of static
offset is that this mirrors how the new ClockSyncManager shifts
timestamps. This makes the follow-up patch to use ClockSyncManager for
timestamp shifting much simpler.

BUG=catapult:#1606

Review URL: https://codereview.chromium.org/1840063004