Refactor ResponsivenessMetric to use diagnostic values
Currently, ResponsivenessMetric adds many Values to the ValueList for the model:
the overall responsiveness score,
responsiveness scores for each R/A/L,
and throughput and smoothness for each A.

However, this is not compatible with how we are planning to use it.

This CL refactors ResponsivenessMetric to add only the overall responsiveness
score to the ValueList, with values for the per-UE scores in the overall value's
diagnostics.

BUG=catapult:#2183

Review URL: https://codereview.chromium.org/1849433004