Rename systrace to run_systrace and agents to tracing_agents.
This is part of the refactoring of systrace to support the new clock
sync design. Renaming systrace to run_systrace helps to reduce confusion
caused by the existence of file systrace in module also named systrace.
Renaming agents to tracing_agents just adds some more clarity. Note
that we also change __init__.py so that the importer can find the new files
without needing to have them in your PYTHONPATH (note that systrace itself
still needs to be in the PYTHONPATH to find the __init__.py file)

Review URL: https://codereview.chromium.org/1858003002