[clock sync] Add Chrome clock domains to the list of valid domains
This fixes a bug that's currently causing the import to fail when a
BattOr trace is recorded in Chrome Canary. I thought that
clock_sync_test.html was testing this case, but because there were no
clock sync markers in the trace event traces, they were never calling
the ClockSyncManager and thus never executing the error checking code.

This CL:

  1) Fixes clock_sync_test.html so that it fails without the Chrome
     clock domains, reproducing the current problem in Canary.
  2) Adds the Chrome clock domains so that the tests don't fail anymore,
     fixing the bug and making the tests pass.

BUG=catapult:#2221

Review URL: https://codereview.chromium.org/1867463004