Redefine ChromeRendererHelper.isRenderProcess
Currently, a process is identified as a renderer process if it has both a
CrRendererMain thread and a Compositor thread.

However, sometimes the 'cc' category is not enabled, so the Compositor thread is
not traced.
Conversely, sometimes 'cc' is the only category that is traced, so there is no
CrRendererMain thread.

This CL redefines isRenderProcess to return true if there is either a
CrRendererMain or a Compositor process, or both, so that the ChromeRendererHelper is usable.

This breaks the guarantee that ChromeRendererHelper will always have both a
Compositor thread and a main thread, so callers will need to explicitly check
for the thread they want.

BUG=catapult:#2178

Review URL: https://codereview.chromium.org/1839673002