[clock sync] Delete logic for BattOr cross-correlation clock sync
This breaks backwards compatibility with old Android/BattOr traces.
However, there are a very limited number of those in the wild so this
change should have minimal impact.

This is necessary because the cross-correlation clock sync uses clock
sync markers in nonstandard ways that break the new ClockSyncManager.
Specifically, there are multiple clock sync markers all with the same
sync ID of 'battor'.

I also swapped out the old 'battor.zip' test data that used this
nonstandard sync behavior with a new Chrome example that uses a
standard sync.

If for some reason we need this logic in the future, we can probably
find a way to add it back in. I suspect this won't be the case, though.

BUG=catapult:#1606
R=petrcermak@chromium.org

Review URL: https://codereview.chromium.org/1861343004 .