[perf-insights] - Refactor map_single_trace_cmdline.html to return MreResult.
Currently this just dumps the MreResult dict to the console. This CL exposes a function to return the MreResult directly, which main uses as before. This allows us to reuse this entry-point for the trace processor and minimize the amount of glue code needed.

Review URL: https://codereview.chromium.org/1895823004