Replace template usage with explicit string substitution
The base-form element was removed because it didn't have much to do with
forms and wasn't a good place to put code to show toasts. It had some
custom template parsing that was also removed. Replace the usages of it
with more explicit string substitution.

BUG=catapult:#2288

Review-Url: https://codereview.chromium.org/1931803004

R=eakuefner@chromium.org

Review URL: https://codereview.chromium.org/1924983003 .