FrameViewer uses layer list instead of layer tree hierarchy
Slimming paint v2 requires chrome compositor to receive a layer list
instead of layer tree. As a result the tracing of layer_impl will not
be able to know anything about the tree hierarchy.

This CL makes the frame viewer work with both a compositor that
knows hierarchy (stores children value in LayerImpl::AsValueInto) and a
compositor that does not know hierarchy information (stores a collection
of layers in LayerTreeImpl::AsValueInto).

BUG=chromium:605940

Review-Url: https://codereview.chromium.org/1947553002