Adds a new admin endpoint, /delete_test_data
In the past, requests to delete data had been rare, but due to the tbm2
migration there are many more requests than we had seen previously.
Examples from this past week include crbug.com/606361 and
crbug.com/607124.

Adding an admin-protected, unit-tested endpoint for this since we want
to ensure that data deletion is done correctly.

BUG=chromium:606361

Review-Url: https://codereview.chromium.org/1951913002