Add a metrics table side panel
Currently, the only metrics that are visible in trace viewer are a few specific
per-UE system health metrics in single-user-expectation-sub-view.

We need to be able to visualize all properties of all metric values
for any given trace.

This CL adds a simple metrics table side panel.
It's far from perfect, but it's usable.

This is a tiny step towards TBMv2.1 and results2.html.
There will be many follow-up CLs.

BUG=catapult:#2094,catapult:#2035

Review-Url: https://codereview.chromium.org/1927723002