[catapult] - Fix bin/run_tests to run tests.
Looks like the path to py_vulcanize/bin/run_py_tests was incorrect, and as a result test_runner.py:_RunTest bails out (without a failure exit code), and we don't run the rest of the tests.

After fixing this, a lot more tests seem to run, and correctly fails to run py_vulcanize's tests.

Review-Url: https://codereview.chromium.org/1979193005