Pass rangeOfInterest to metrics for the metrics side panel.
Currently, metrics always act on the entire model.
However, in trace-viewer, it is useful to be able to select a part of the model
and ignore the rest. The metric side panel would be more useful if it allowed
users to compute metrics for a selected time range.

This CL updates responsivenessMetric and longTasksMetric as examples of this
new capability, though any metric can optionally take a third "opt_options"
parameter containing "rangeOfInterest". More options could be added to opt_options.

Review-Url: https://codereview.chromium.org/2052593005