Define createDiagnosticSpan()
Currently, Diagnostics can be serialized and deserialized, but not displayed.

This CL defines polymer elements related-value-set-span, generic-diagnostic-span,
and a factory function tr.v.ui.createDiagnosticSpan(diagnostic).

A composition-span is left for a follow-up CL since it may involve remembering D3.

createDiagnosticSpan falls back to creating the spans registered for base classes
when a Diagnostic subclass does not register an element name, as is currently the
case for Composition.

RelatedValueMap will allow metrics and summarization to customize the
analysis-link contents: https://codereview.chromium.org/2045993002

BUG=catapult:#2180

Review-Url: https://codereview.chromium.org/2038683002